+++ R/01-default-functions.R
@@ -3,9 +3,9 @@
 # @param split A character string containing a regular expression to use
 #              as the split point.
 # @param ... Additional arguments passed to `strsplit()`.
-# @return A character vector with the split elements.
+# @return A list of character strings resulting from splitting `x` by `split`.
 simple_split <- function(x, split, ...) {
-  strsplit(x, split, ...)[[1]]
+  as.list(strsplit(x, split, ...)[[1]])
 }
 
 #' Fetch and parse web text from a URL